backend-python/requirements.txt
# Backend Python Requirements

# FastAPI and Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.6

# HTTP Client
httpx==0.27.0
aiohttp>=3.9.0

# Catalog Indexing (Core Dependencies)
sqlalchemy>=2.0.0
sentence-transformers>=2.2.0
chromadb>=0.5.15
rank-bm25==0.2.2

# LLM Integration (TODO: Uncomment when implementing)
# openai>=1.0.0

# Utilities
tqdm>=4.65.0
python-dotenv==1.0.0
pandas>=2.0.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio==0.21.1
